refactor(pages): split edit-challenge and habits pages
- challenge/edit/[id]/page.tsx 346 -> 116: extract <EditChallengeForm> to
  src/components/EditChallengeForm.tsx (owns form state, validation, submit).
- habits/page.tsx 297 -> 179: extract useHabitActions hook to consolidate
  the five mutation handlers (check-in/skip/undo/stop/delete) which all
  shared the same try/reload/log pattern.

Both pages now fit under the 180-LOC page target. Behaviour unchanged;
all 1802 tests still pass.
