+++ src/app/habits/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 /**
  * Habits Management Page
  *
@@ -5,36 +7,22 @@
  * Shows active habits, completed habits, and overall statistics.
  */
 
-'use client';
+import { LightBulbIcon } from '@primer/octicons-react';
+import { Banner, Heading, Spinner, Stack, Text } from '@primer/react';
+import dynamic from 'next/dynamic';
+import { useCallback, useEffect, useState } from 'react';
 
 import { AppHeader } from '@/components/AppHeader';
-import { HabitStatsSection } from '@/components/Habits/habit-stats-section';
 import { HabitListSection } from '@/components/Habits/habit-list-section';
+import { HabitStatsSection } from '@/components/Habits/habit-stats-section';
 import { ProfileNav } from '@/components/ProfileNav';
 import { useBreadcrumb } from '@/contexts/breadcrumb-context';
+import { useHabitActions } from '@/hooks/use-habit-actions';
 import { habitStore } from '@/lib/habits';
-import {
-  checkInHabit,
-  skipHabitDay,
-  undoCheckIn,
-} from '@/lib/habits/state-machine';
 import type { HabitWithHistory } from '@/lib/habits/types';
 import { logger } from '@/lib/logger';
-import {
-  LightBulbIcon,
-} from '@primer/octicons-react';
-import {
-  Banner,
-  Heading,
-  Spinner,
-  Stack,
-  Text,
-  useConfirm,
-} from '@primer/react';
-import { useCallback, useEffect, useState } from 'react';
-import styles from './habits.module.css';
 import layoutStyles from '@/styles/two-column-layout.module.css';
-import dynamic from 'next/dynamic';
+import styles from './habits.module.css';
 
 // Lazy-load dialog components — they are only needed on first user interaction,
 // so we defer their JS chunk until the user opens a dialog.
@@ -57,8 +45,6 @@ export default function HabitsPage() {
   const [abandonedHabits, setAbandonedHabits] = useState<HabitWithHistory[]>([]);
   const [isCreateDialogOpen, setIsCreateDialogOpen] = useState(false);
   const [editingHabit, setEditingHabit] = useState<HabitWithHistory | null>(null);
-  const [actionError, setActionError] = useState<string | null>(null);
-  const confirm = useConfirm();
 
   const loadHabits = useCallback(async () => {
     try {
@@ -79,105 +65,10 @@ export default function HabitsPage() {
     }
   }, []);
 
-  useEffect(() => {
-    loadHabits();
-  }, [loadHabits]);
+  useEffect(() => { loadHabits(); }, [loadHabits]);
 
-  const handleCheckIn = useCallback(
-    async (habit: HabitWithHistory, value: number | boolean) => {
-      setActionError(null);
-      try {
-        const updated = checkInHabit(habit, value);
-        await habitStore.update(updated);
-        await loadHabits();
-      } catch (error) {
-        logger.error('Failed to check in', { error, habitId: habit.id }, 'HabitsPage');
-        setActionError(error instanceof Error ? error.message : 'Action failed. Please try again.');
-      }
-    },
-    [loadHabits]
-  );
+  const actions = useHabitActions(loadHabits);
 
-  const handleSkip = useCallback(
-    async (habit: HabitWithHistory) => {
-      setActionError(null);
-      try {
-        const updated = skipHabitDay(habit);
-        await habitStore.update(updated);
-        await loadHabits();
-      } catch (error) {
-        logger.error('Failed to skip', { error, habitId: habit.id }, 'HabitsPage');
-        setActionError(error instanceof Error ? error.message : 'Action failed. Please try again.');
-      }
-    },
-    [loadHabits]
-  );
-
-  const handleUndo = useCallback(
-    async (habit: HabitWithHistory) => {
-      setActionError(null);
-      try {
-        const updated = undoCheckIn(habit);
-        await habitStore.update(updated);
-        await loadHabits();
-      } catch (error) {
-        logger.error('Failed to undo check-in', { error, habitId: habit.id }, 'HabitsPage');
-        setActionError(error instanceof Error ? error.message : 'Action failed. Please try again.');
-      }
-    },
-    [loadHabits]
-  );
-
-  const handleDelete = useCallback(
-    async (habit: HabitWithHistory) => {
-      setActionError(null);
-      const confirmed = await confirm({
-        title: 'Delete Habit',
-        content: `Are you sure you want to delete "${habit.title}"? This action cannot be undone.`,
-        confirmButtonContent: 'Delete',
-        confirmButtonType: 'danger',
-      });
-
-      if (confirmed) {
-        try {
-          await habitStore.delete(habit.id);
-          await loadHabits();
-          logger.info('Habit deleted', { habitId: habit.id }, 'HabitsPage');
-        } catch (error) {
-          logger.error('Failed to delete habit', { error, habitId: habit.id }, 'HabitsPage');
-          setActionError(error instanceof Error ? error.message : 'Action failed. Please try again.');
-        }
-      }
-    },
-    [confirm, loadHabits]
-  );
-
-  const handleStop = useCallback(
-    async (habit: HabitWithHistory) => {
-      setActionError(null);
-      const confirmed = await confirm({
-        title: 'Stop Habit',
-        content: `Are you sure you want to stop "${habit.title}"? You can always view it in the Stopped Habits section.`,
-        confirmButtonContent: 'Stop Habit',
-        confirmButtonType: 'danger',
-      });
-
-      if (confirmed) {
-        try {
-          const updated: HabitWithHistory = { ...habit, state: 'abandoned' };
-          await habitStore.update(updated);
-          await loadHabits();
-          logger.info('Habit stopped', { habitId: habit.id }, 'HabitsPage');
-        } catch (error) {
-          logger.error('Failed to stop habit', { error, habitId: habit.id }, 'HabitsPage');
-          setActionError(error instanceof Error ? error.message : 'Action failed. Please try again.');
-        }
-      }
-    },
-    [confirm, loadHabits]
-  );
-
-  // Calculate statistics
   const totalCheckIns = [...activeHabits, ...completedHabits].reduce(
     (sum, h) => sum + h.checkIns.length,
     0
@@ -192,9 +83,7 @@ export default function HabitsPage() {
         <main className={layoutStyles.main}>
           <aside className={layoutStyles.sidebar}>
             <ProfileNav />
-            <div className={layoutStyles.sidebarCard}>
-              <Spinner size="medium" />
-            </div>
+            <div className={layoutStyles.sidebarCard}><Spinner size="medium" /></div>
           </aside>
           <div className={styles.content}>
             <Stack align="center" justify="center" style={{ flex: 1, padding: '48px' }}>
@@ -212,7 +101,6 @@ export default function HabitsPage() {
       <AppHeader />
 
       <main className={layoutStyles.main}>
-        {/* Left Sidebar */}
         <aside className={layoutStyles.sidebar}>
           <ProfileNav />
 
@@ -234,9 +122,7 @@ export default function HabitsPage() {
           </div>
         </aside>
 
-        {/* Main Content */}
         <div className={styles.content}>
-          {/* Header */}
           <div className={styles.header}>
             <Heading as="h1">My Habits</Heading>
             <Text as="p" style={{ color: 'var(--fgColor-muted)', marginTop: '4px' }}>
@@ -245,18 +131,14 @@ export default function HabitsPage() {
           </div>
 
           {loadError && (
-            <Banner
-              title="Failed to load habits"
-              description={loadError}
-              variant="critical"
-            />
+            <Banner title="Failed to load habits" description={loadError} variant="critical" />
           )}
-          {actionError && (
+          {actions.actionError && (
             <Banner
               title="Action failed"
-              description={actionError}
+              description={actions.actionError}
               variant="critical"
-              onDismiss={() => setActionError(null)}
+              onDismiss={actions.dismissError}
             />
           )}
 
@@ -264,12 +146,12 @@ export default function HabitsPage() {
             activeHabits={activeHabits}
             completedHabits={completedHabits}
             abandonedHabits={abandonedHabits}
-            onCheckIn={handleCheckIn}
-            onSkip={handleSkip}
-            onUndo={handleUndo}
+            onCheckIn={actions.checkIn}
+            onSkip={actions.skip}
+            onUndo={actions.undo}
             onEdit={setEditingHabit}
-            onStop={handleStop}
-            onDelete={handleDelete}
+            onStop={actions.stop}
+            onDelete={actions.remove}
             onNewHabitClick={() => setIsCreateDialogOpen(true)}
           />
         </div>
